+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
-* Set min SDK version to `2.9.0`
-* Changed the support version range of the analyzer to `>=0.40.0 <0.42.0`
-* Changed the support version range of the analyzer_plugin to `>=0.3.0 <0.5.0`
+* Set min `SDK` version to `2.9.0`
+* Set min `args` 1.6.0
+* Changed the support version range of the `analyzer` to `>=0.40.0 <0.42.0`
+* Changed the support version range of the `analyzer_plugin` to `>=0.3.0 <0.5.0`
 * Remove obsolete rule `prefer-trailing-comma-for-collection`
+* Rename cli arguments:
+  * `ignore-files` to `exclude`
+  * `maximum-nesting` to `maximum-nesting-level`
+  * `number-of-arguments` to `number-of-parameters`
+* Update README
 
 ## 2.4.1